Extract Enums to data_struct.py
