+++ magia/bundle.py
@@ -3,9 +3,8 @@ from __future__ import annotations
 import enum
 from copy import deepcopy
 
-from .constants import SignalType
 from .core import Signal, SignalConfig
-from .data_struct import SignalDict
+from .data_struct import SignalDict, SignalType
 from .io_signal import Input, Output
 from .module import Instance, IOPorts
